+++ src/Framework/AbstractFactory.php
@@ -48,12 +48,14 @@ abstract class AbstractFactory
         $resolver = (new ProviderResolver())->resolve($this);
         $resolver?->provideModuleDependencies($container);
 
-        // Temporal solution to keep BC with the AbstractDependencyProvider
-        $dpResolver = (new DependencyProviderResolver())->resolve($this);
-        $dpResolver?->provideModuleDependencies($container);
+        {
+            // Temporal solution to keep BC with the AbstractDependencyProvider
+            $dpResolver = (new DependencyProviderResolver())->resolve($this);
+            $dpResolver?->provideModuleDependencies($container);
 
-        if (!$resolver instanceof AbstractProvider && !$dpResolver instanceof AbstractProvider) {
-            throw new ProviderNotFoundException(static::class);
+            if (!$resolver instanceof AbstractProvider && !$dpResolver instanceof AbstractProvider) {
+                throw new ProviderNotFoundException(static::class);
+            }
         }
 
         return $container;