+++ tests/Feature/Framework/BindingsInsideProvider/Module/Domain/Greeter/Team.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Framework\BindingsInsideProvider\Module\Domain\Greeter;
+
+final class Team
+{
+    public function getNames(): string
+    {
+        return 'Chemaclass & Jesus';
+    }
+}